update: ui layout cleanup for pro users
